Fix device list of loaded executable in PJRT plugin for multiple GPUs (#19369)
It closes #19366, and blocks #19279.

After this PR, `ClientOptions::Compile` will first check the device
assignment in the compile options, and then return the corresponding
device list with the loaded executable.

To achieve this, we introduce protobuf via `FetchContent` in this PR,
which is scoped to the PJRT plugin. Compile options will be passed by
the PJRT client encoded in protobuf, and in this plugin we decode it
first and then retrieve some interesting fields.

ci-exactly: build_packages, test_pjrt

---------

Signed-off-by: PragmaTwice <twice@apache.org>
Co-authored-by: Scott Todd <scott.todd0@gmail.com>