Drop unnecessary arg-ids in `llvm::formatv` (#19426)
`llvm::formatv` supports (approximately?) the same format syntax as
`std::format` described
[here](https://en.cppreference.com/w/cpp/utility/format/format). The
general pattern is `{arg-id}` or `{arg-id:format-spec}` where `arg-id`
is optional and may be omitted to default to consuming the args in the
order in which they are provided. So the only reason to ever specify
`arg-id` is to consume the same arg repeatedly or in a different order.

It turns out that 100% of the call sites were consuming args in the
order in which they were specified, so the arg-ids were unnecessary.

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>