[DT][VMVX] Implement VMVXEncodingLayoutAttr. (#19403)
The revision implements the VMVX encoding layout attribute for device
code with the new IREE::CPU dialect.

Some additional dialect registration is needed because it could create
the VMVXEncodingLayoutAttr attribute in the pass pipeline; the dialect
needs to be loaded. The main refactoring changes are:

- Move chooseMatmulTile to IREECodegen/Utils. We can internalize the
  method once other CPU materialization logic is moved to their own
  attribute implementation.
- Move the enumerateMatmulTilesVMVX method to
  CPUEncodingExternalModels.cpp. It is only used by VMVX implementation.

On the Codegen utilitiy side (i.e., Codegen/Utils.[h|cpp]), the revision
adapts config query methods to use `Attribute`, and share the logic with
DictionaryAttr type. Previously, the configuration is wrapped into
IREE::HAL::ExecutableTarget attribute. Now we have DictionaryAttr
variants and we do not want to duplicate the implementation. Thus, the
functions take Attribute input and handle the cases in the
implementation.

On the data-tiling encoding materialization pass side, it creates the
VMVXEncodingLayoutAttr attribute (with the original target
configuration).

Note that the target configuration is an optional parameter and it is
expected to be used within pass scope, but not the final IR output.

No additional tests because they are covered by
vmvx_materialize_encoding.mlir.

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>