Use LLVMGPUTileandFuse instead of LLVMGPUVectorize for convolutions (#19469)
With this PR for convs that are not picked by VectorDistribute or
TileAndFuse via IGEMM, we default lower them with TileAndFuse instead of
using Vectorize pipeline. There doesnt seem to be a major performance
impact in testing done with iree-kernel-benchmark as shown
[here](https://docs.google.com/spreadsheets/d/1WaJ1ELhwdo1wFvNiKbdoddSncSt2_UsbvrTdSObNaAo/edit?gid=0#gid=0)
and we can always look into improving the heuristics if performance is a
problem.
Fixes https://github.com/iree-org/iree/issues/19478

---------

Signed-off-by: Nirvedh <nirvedh@gmail.com>