[PJRT] Fix compile error while tracing is enabled (#19485)
While trace is enabled in the PJRT plugin, i.e. cmake option
`-DIREE_ENABLE_RUNTIME_TRACING=ON` is passed, the compilation will abort
due to a wrong use of `IREE_TRACE_SCOPE_NAMED`, where an undefined
identifier is passed as a name. And it should be a string literal
instead, as in other places.

ci-exactly: build_packages, test_pjrt

Signed-off-by: PragmaTwice <twice@apache.org>