Adding iree_io_file_handle_create/iree_io_file_handle_open. (#19510)
These utilities are optional but an easy way to get a file handle that
is compatible with the rest of IREE. This largely replaces the
need for the existing file_io utilities but cleanup/unification is left
for future changes.

Command line tools now default to --parameter_mode=file though users can
force mmap or preload. There's some flags that are available on the open
that may help or hurt performance and the defaults are provisional.