GPU Data-tiled multi-mma: subgroup dimensions should be outer (#19521)
This was already the idea, but there was an accidental exception: in the
accumulator tensor, if there was both a `unroll_m` dimension and
`subgroup_n` dimension, then the `subgroup_n` dimension wasn't on the
outside of `unroll_m` as it was meant to be.

Noticed this when it required corresponding strides in the ukernel.

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>