[Flow] Improve DeduplicateExecutables bucketing (#19601)
This change uses the result types as a part of the hash when grouping
ops. This vastly improves the performance of this pass when there are
several similar objects that consist of ops with the same names but
differ in the number/type of results. However, this may increase the
overhead of hashing when bucketing isn't effective.

Although a sample size of one, I found that for 405b tp8, the number of
buckets went from 35 -> 140. This brought the time of this pass from a
few minutes to several seconds.

Signed-off-by: Ian Wood <ianwood2024@u.northwestern.edu>