[Dispach] Clone chain of ops into dispatch (#19723)
This change modifies `cloneProducersToRegion` to iteratively clone
producers into the region. For example if ther is a chain A -> B ->
Dispatch, where A and B are cloneable. `A` is only detected as a
clonable op after `B` has been cloned.


Also traverses backwards to delete trivially dead ops so that consumers
get erased first.

Signed-off-by: Ian Wood <ianwood2024@u.northwestern.edu>