[DT][GPU] Permute cross-thread dims of TileSwizzle to outermost (#19734)
This PR adds a new constraint on the generated TileSwizzles during GPU
encoding materialization. The new constraint is that all `CrossThread`
dimensions in the ACC layout must come from the outermost dimensions
within their reassociation groups of the swizzle's expand_shape. For
example, consider the following TileSwizzle for an ACC layout generated
without this constraint:

```
expandShape = [[8, {4}, 4], [{4}, 2, {16}]],
permutation = [3, 0, 4, 1, 5, 2]
```

`CrossThread` dimensions are denoted with braces {}, and this
TileSwizzle does not follow the new constrain because the first `{4}`
dim of reassociation group 0 and the `{16}` dim of reassociation group 1
are not outermost. After adding the new constraint, this swizzle would
look like the following:

```
expandShape = [[{4}, 8, 4], [{4}, {16}, 2]],
permutation = [3, 1, 5, 0, 4, 2]
```

Now, all `CrossThread` dimensions are outermost within their
reassociation groups. The permutation has been adjusted accordingly so
that the result shape of the swizzled tile remains the same in both
cases.

The LHS and RHS swizzle layouts also have to be adjusted to match the
new ACC layout, but the CrossThread dimensions are not necessarily the
same between corresponding M and N tiles of the ACC swizzle and LHS/RHS
swizzles. Because of this, the LHS and RHS (currently only LHS needs
this) swizzle shapes must be expanded to match the dimensionality of the
ACC layout. This is the reason why some of the LHS layouts have
additional expansion after this PR.

The reason for adding this constraint is so that the swizzle operations
of unset_encoding operations are able to be fused into the thread loop
of their data tiled multi_mma operation. This constraint makes the
fusion possible by forcing the slices that are held by a thread at the
end of the multi_mma computation to be contiguous in the linear layout
tensor within each reassociation group. This matters because we need to
fuse the collapse_shape op of the unset_encoding into the thread loop,
which is only possible when the written slice is contiguous in the
result for each reassociation group.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>