Port compiler and iree-dialects to nanobind (#19790)
Upstream MLIR ported its Python bindings from pybind11 to nanobind and
so do we. With this we have a single library for the binding across the
compiler, iree-dialects and the runtime.

The macOS build is expected to still fail, see #19591, but as upstream
switched to build the bindings for the internal dialects with nanobind
it does anyway without reverts carried forward in our LLVM fork.