[LLVMGPU] Fix lowering of functions that don't use all bindings (#19773)
Runtime changes in [some previous PR, ask Ben] mean that now, GPU
kernels are passed one pointer for each binding in the pipeline layout,
whether or not they are used. When the previous behavior, which was to
only pass in the needed pointers one after another, was removed, the GPU
code was not updated to mach.

This PR updates the conversion from func.func to llvm.func to use one
pointer per binding. It also moves the setting of attributes like
noundef or nonnull into the function conversion, instead of making the
lowerigs for hal.interface.binding.subspan and
hal.interface.constant.load reduntantly add those attributes.