Add pattern to convert generic conv ops to IGEMM (#19798)
This PR removes the named op patterns to convert convs to IGEMM and
replaces them with a generic pattern that works for all supported convs.
A new utility function that populates the shared details required for
setting lowering config and doing the IGEMM computation is added.
The PR is currently using a default true flag
`iree-gpu-use-tile-and-fuse-generic-convolution` . The idea is that
since a lot more convolutions will go down the IGEMM path with this PR
if any of them run into issues we can turn the flag off by default
rather then needing to revert the whole PR. If after some time we find
that there are no issues then we can drop the flag and have this
happening always.

---------

Signed-off-by: Nirvedh Meshram <nirvedh@gmail.com>