[GPU] Use tile and fuse for matmul after vector distribute by default (#19884)
Currently some efforts such as
https://github.com/iree-org/iree/pull/19854 and
https://github.com/iree-org/iree/pull/19520 are ongoing to make the Tile
and Fuse matmul pipeline on by default. However, these efforts are still
WIP in achieving exact parity with the current default of Vector
Distribute in all use cases.
This PR in the time being tries Tile and Fuse after Vector Distribute so
that we can get the benefits of tile and fuse such as handling unaligned
to intrinsic shape while leaving the shapes that vector distribute
handles untouched.
Fixes :  https://github.com/iree-org/iree/issues/19864
Fixes: https://github.com/iree-org/iree/issues/19855

---------

Signed-off-by: Nirvedh Meshram <nirvedh@gmail.com>