Integrate LLVM at 73f11ac (#19941)
We had previously cherry-picked
https://github.com/llvm/llvm-project/commit/73f11ac17d50f0585d03a880e756ff13c321bfd0
in https://github.com/iree-org/iree/pull/19939.

Now we're integrating up to that commit, so it's no longer a
cherry-pick.

Reverting https://github.com/llvm/llvm-project/pull/125789 because it
breaks TorchToTosa, in torch-mlir. We will need to wait for this to be
resolved in torch-mlir, then simultaneously bump torch-mlir and drop the
revert.

Chery-pick a Bazel fix:
https://github.com/llvm/llvm-project/commit/4df287a171b63514a5028b85272fcc1b89555ee4

---------

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>